+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/). Release checklist in [deploy.md](docs/specs/deploy.md).
 
+## [0.6.2] - 2026-04-13
+- Fix issues with deployed Tauri on Win and Mac (Linux is working great!)
+
+## [0.6.1] - 2026-04-13
+- Fix missing Tauri update permissions.
+
+## [0.6.0] - 2026-04-13
+- Standalone: fix some issues with node sidecar.
+- Standalone: app-rendered title bar.
+
 ## [0.5.2] - 2026-04-10
 - Codex fixes.